[Sneha Agarwal]Add. Program to Calculate the power from base and exponent

+++ core-java-practice/gcr-codebase/PowerCalculation.java
@@ -1,5 +1,4 @@
-
-import java.util.Scanner;
+import java.util.*;
 
 public class PowerCalculation {
     public static void main(String[] args) {
@@ -7,13 +6,13 @@ public class PowerCalculation {
         Scanner sc=new Scanner(System.in);
 
         System.out.println("Enter the base");
-        int base=sc.nextInt(); //take base as input
+        double base=sc.nextDouble(); //take base as input
 
         System.out.println("Enter the exponent");
-        int exponent=sc.nextInt(); //take exponent as input
+        double exponent=sc.nextDouble(); //take exponent as input
 
         //calculate the power
-        double power=Math.pow(base, exponent);
+        double power=Math.pow(base,exponent);
 
         System.out.println("The solution of power calculation is: " + power);
